Move rotator to vault package
