+++ cli/login.go
@@ -7,8 +7,6 @@ import (
 	"log"
 	"net/http"
 	"net/url"
-	"regexp"
-	"strings"
 	"time"
 
 	"github.com/99designs/aws-vault/prompt"
@@ -17,7 +15,6 @@ import (
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/credentials"
 	"github.com/aws/aws-sdk-go/aws/session"
-	"github.com/aws/aws-sdk-go/service/iam"
 	"github.com/aws/aws-sdk-go/service/sts"
 	"github.com/skratchdot/open-golang/open"
 	"gopkg.in/alecthomas/kingpin.v2"
@@ -201,7 +198,7 @@ func getFederationToken(creds credentials.Value, d time.Duration) (*sts.Credenti
 	})
 	client := sts.New(sess)
 
-	currentUsername, err := getCurrentUserName(sess)
+	currentUsername, err := vault.GetUsernameFromSession(sess)
 	if err != nil {
 		return nil, err
 	}
@@ -218,33 +215,4 @@ func getFederationToken(creds credentials.Value, d time.Duration) (*sts.Credenti
 	}
 
 	return resp.Credentials, nil
-}
-
-var (
-	getUserErrorRegexp = regexp.MustCompile(`^AccessDenied: User: arn:aws:iam::(\d+)(.+) is not`)
-)
-
-func getCurrentUserName(sess *session.Session) (string, error) {
-	client := iam.New(sess)
-
-	resp, err := client.GetUser(&iam.GetUserInput{})
-	if err != nil {
-		matches := getUserErrorRegexp.FindAllStringSubmatch(err.Error(), -1)
-		if len(matches) > 0 {
-			log.Printf("%#v", matches)
-		}
-
-		return "", err
-	}
-
-	if resp.User.UserName != nil {
-		return *resp.User.UserName, nil
-	}
-
-	if resp.User.Arn != nil {
-		arnParts := strings.Split(*resp.User.Arn, ":")
-		return arnParts[len(arnParts)-1], nil
-	}
-
-	return "", fmt.Errorf("Couldn't determine current username")
 }
